+++ src/faq.njk
@@ -1,7 +1,7 @@
 ---
 layout: layout.njk
 pagination:
-  data: allContent.faqItems
+  data: data.faqItems
   size: 1
   alias: item
 permalink: "/faq/{{ item.category }}/{{ item.filename | replace('.md', '') }}/"
