+++ src/pending-guidance-index.njk
@@ -12,7 +12,7 @@ permalink: "/pending-guidance/"
   <section class="category-section">
     <h2 class="category-title">Pending Guidance</h2>
     <ul class="question-list">
-      {% for item in allContent.guidance %}
+      {% for item in data.guidance %}
         <li class="question-item">
           <a href="{{ item.permalink }}" class="question-link">
             <span class="question-text">
