+++ src/pending-guidance.njk
@@ -1,7 +1,7 @@
 ---
 layout: layout.njk
 pagination:
-  data: allContent.guidance
+  data: data.guidance
   size: 1
   alias: item
 permalink: "/pending-guidance/{{ item.filename | replace('.md', '') }}/"
